Started trying to make delete and edit buttons

+++ Cofano-C/src/main/java/nl/utwente/di14/Cofano_C/resources/ShipsResource.java
@@ -164,16 +164,17 @@ public class ShipsResource {
 	@Path("/{shipId}")
 	public void deletShip(@PathParam("shipId") int shipId, @Context HttpServletRequest request) {
 		Tables.start();
-
-		String query ="DELETE FROM ship WHERE sid = ?";
-		try {
-			PreparedStatement statement = Tables.getCon().prepareStatement(query);
-			statement.setLong(1, shipId);
-			statement.executeUpdate();
-		} catch (SQLException e) {
-			System.err.println("Was not able to delete APP");
-			System.err.println(e.getSQLState());
-			e.printStackTrace();
+		if(request.getSession().getAttribute("userEmail")!=null) {
+			String query ="DELETE FROM ship WHERE sid = ?";
+			try {
+				PreparedStatement statement = Tables.getCon().prepareStatement(query);
+				statement.setLong(1, shipId);
+				statement.executeUpdate();
+			} catch (SQLException e) {
+				System.err.println("Was not able to delete APP");
+				System.err.println(e.getSQLState());
+				e.printStackTrace();
+			}
 		}
 	}
 
@@ -183,24 +184,26 @@ public class ShipsResource {
 	@Produces(MediaType.APPLICATION_JSON)
 	public Ship getShip(@PathParam("shipId") int shipId, @Context HttpServletRequest request) {
 		Ship ship = new Ship();
-		String query = "SELECT * FROM ship WHERE sid = ?";
-
-		try {
-			PreparedStatement statement = Tables.getCon().prepareStatement(query);
-			statement.setInt(1, shipId);
-			ResultSet resultSet = statement.executeQuery();
-
-			while(resultSet.next()) {
-				ship = new Ship();
-				ship.setName(resultSet.getString(3));
-				ship.setImo(resultSet.getString(2));
-				ship.setID(resultSet.getInt(1));
-				ship.setDepth(resultSet.getBigDecimal(6));
-				ship.setCallsign(resultSet.getString(4));
-				ship.setMmsi(resultSet.getString(5));
+		if(request.getSession().getAttribute("userEmail")!=null) {
+			String query = "SELECT * FROM ship WHERE sid = ?";
+	
+			try {
+				PreparedStatement statement = Tables.getCon().prepareStatement(query);
+				statement.setInt(1, shipId);
+				ResultSet resultSet = statement.executeQuery();
+	
+				while(resultSet.next()) {
+					ship = new Ship();
+					ship.setName(resultSet.getString(3));
+					ship.setImo(resultSet.getString(2));
+					ship.setID(resultSet.getInt(1));
+					ship.setDepth(resultSet.getBigDecimal(6));
+					ship.setCallsign(resultSet.getString(4));
+					ship.setMmsi(resultSet.getString(5));
+				}
+			} catch (SQLException e) {
+				e.printStackTrace();
 			}
-		} catch (SQLException e) {
-			e.printStackTrace();
 		}
 
 
@@ -210,22 +213,24 @@ public class ShipsResource {
 	@PUT
 	@Path("/{shipId}")
 	@Consumes(MediaType.APPLICATION_JSON)
-	public void updateApp(@PathParam("shipId") int shipId, Ship ship) {
+	public void updateApp(@PathParam("shipId") int shipId, Ship ship,@Context HttpServletRequest request) {
 		System.out.println("JOOOOOSOSKSOKSOSKSOKS");
-
-		String query = "UPDATE ship SET imo = ?, name = ?, callsign = ?, mmsi = ?, ship_depth = ? WHERE sid = ?";
-		try {
-			PreparedStatement statement = Tables.getCon().prepareStatement(query);
-			statement.setString(1, ship.getImo());
-			statement.setString(2, ship.getName());
-			statement.setString(3, ship.getCallsign());
-			statement.setString(4, ship.getMmsi());
-			statement.setBigDecimal(5, ship.getDepth());
-			statement.setInt(6, shipId);
-			statement.executeQuery();
-
-		} catch (SQLException e) {
-			e.printStackTrace();
+		if(request.getSession().getAttribute("userEmail")!=null) {
+	
+			String query = "UPDATE ship SET imo = ?, name = ?, callsign = ?, mmsi = ?, ship_depth = ? WHERE sid = ?";
+			try {
+				PreparedStatement statement = Tables.getCon().prepareStatement(query);
+				statement.setString(1, ship.getImo());
+				statement.setString(2, ship.getName());
+				statement.setString(3, ship.getCallsign());
+				statement.setString(4, ship.getMmsi());
+				statement.setBigDecimal(5, ship.getDepth());
+				statement.setInt(6, shipId);
+				statement.executeQuery();
+	
+			} catch (SQLException e) {
+				e.printStackTrace();
+			}
 		}
 
 	}
